requirements.txt
audioread==3.0.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
opencv-python==4.8.1.78
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
proglog==0.1.10
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
requests==2.31.0
resampy==0.4.3
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.1.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
threadpoolctl==3.3.0
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
